Module chain_actions::node

This module contains the NodeActions trait defining action which relate to HOPR node itself.

An implementation of this trait is added to ChainActions which realizes the redemption operations via ActionQueue.

There are 3 functions that can be used to redeem tickets in the NodeActions trait:

withdraw

announce

register_safe_by_node

All necessary pre-requisites are checked by the implementation before the respective Action is submitted to the ActionQueue.

Traits

NodeActions

Contains all on-chain calls specific to HOPR node itself.