Module chain_actions::payload

Module defining various Ethereum transaction payload generators for the actions.

This module defines the basic PayloadGenerator trait that describes how an action is translated into a [TypedTransaction] that can be submitted on-chain (via an RPC provider) using a TransactionExecutor.

There are two main implementations:

BasicPayloadGenerator which implements generation of a direct EIP1559 transaction payload. This is currently not used by a HOPR node.

SafePayloadGenerator which implements generation of a payload that embeds the transaction data into the SAFE transaction. This is currently the main mode of HOPR node operation.

Structs

BasicPayloadGenerator

Generates transaction payloads that do not use Safe-compliant ABI

SafePayloadGenerator

Payload generator that generates Safe-compliant ABI

Constants

DEFAULT_TX_GAS

Traits

PayloadGenerator

Trait for various implementations of generators of common on-chain transaction payloads.

Functions

convert_acknowledged_ticket

Convert off-chain representation of acknowledged ticket to representation that the smart contract understands

convert_vrf_parameters

Converts off-chain representation of VRF parameters into a representation that the smart contract understands