chain_actions::payload
Function convert_acknowledged_ticket

pub fn convert_acknowledged_ticket(
    off_chain: &RedeemableTicket,
) -> Result<RedeemableTicket>

Convert off-chain representation of acknowledged ticket to representation that the smart contract understands

Not implemented using From trait because logic fits better here