Function chain_actions::payload::convert_vrf_parameters

pub fn convert_vrf_parameters(
    off_chain: &VrfParameters,
    signer: &Address,
    ticket_hash: &Hash,
    domain_separator: &Hash
) -> Vrfparameters

Converts off-chain representation of VRF parameters into a representation that the smart contract understands

Not implemented using From trait because logic fits better here