Module chain_actions::payload

Structs

BasicPayloadGenerator

Generates transaction payloads that do not use Safe-compliant ABI

SafePayloadGenerator

Payload generator that generates Safe-compliant ABI

Enums

Operation

Constants

DEFAULT_TX_GAS

Traits

PayloadGenerator

Trait for various implementations of generators of common on-chain transaction payloads.

Functions

convert_acknowledged_ticket

Convert off-chain representation of acknowledged ticket to representation that the smart contract understands

convert_vrf_parameters

Converts off-chain representation of VRF parameters into a representation that the smart contract understands